+++ example/tiger/manual/tiger.spoofax/src/main/java/mb/tiger/spoofax/task/TigerGetSourceFiles.java
@@ -1,46 +1,35 @@
-package {{baseGetSourceFilesTaskDef.packageId}};
+package mb.tiger.spoofax.task;
 
 import mb.common.util.ListView;
 import mb.pie.api.ExecContext;
 import mb.pie.api.TaskDef;
-import mb.pie.api.stamp.resource.ResourceStampers;
 import mb.resource.hierarchical.HierarchicalResource;
 import mb.resource.hierarchical.ResourcePath;
 import mb.resource.hierarchical.match.ResourceMatcher;
 import mb.resource.hierarchical.match.path.PathMatcher;
 import mb.resource.hierarchical.walk.ResourceWalker;
+import mb.tiger.spoofax.TigerScope;
 
 import javax.inject.Inject;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-@{{adapterProject.scope.qualifiedId}}
-public class {{baseGetSourceFilesTaskDef.id}} implements TaskDef<ResourcePath, ListView<ResourcePath>> {
+@TigerScope
+public class TigerGetSourceFiles implements TaskDef<ResourcePath, ListView<ResourcePath>> {
     private static final ResourceWalker WALKER = ResourceWalker.ofNoHidden();
-    private static final ResourceMatcher MATCHER = ResourceMatcher.ofPath(
-        PathMatcher.ofExtensions({{#shared.fileExtensions}}"{{this}}"{{^-last}}, {{/-last}}{{/shared.fileExtensions}})
-    ).and(ResourceMatcher.ofFile());
-
-    private final {{classLoaderResourcesInput.classLoaderResources.qualifiedId}} classLoaderResources;
+    private static final ResourceMatcher MATCHER = ResourceMatcher.ofPath(PathMatcher.ofExtensions("tig")).and(ResourceMatcher.ofFile());
 
     @Inject
-    public {{baseGetSourceFilesTaskDef.id}}(
-        {{classLoaderResourcesInput.classLoaderResources.qualifiedId}} classLoaderResources
-    ) {
-        this.classLoaderResources = classLoaderResources;
-    }
+    public TigerGetSourceFiles() {}
 
     @Override public String getId() {
-        return "{{baseGetSourceFilesTaskDef.qualifiedId}}";
+        return getClass().getName();
     }
 
     @Override public ListView<ResourcePath> exec(ExecContext context, ResourcePath input) throws Exception {
-        context.require(classLoaderResources.tryGetAsLocalResource(getClass()), ResourceStampers.hashFile());
-
         final HierarchicalResource rootDirectory = context.getHierarchicalResource(input);
         // Require directories recursively, so we re-execute whenever a file is added/removed from a directory.
         rootDirectory.walkForEach(WALKER, ResourceMatcher.ofDirectory(), context::require);
-
         try(final Stream<? extends HierarchicalResource> files = rootDirectory.walk(WALKER, MATCHER)) {
             return ListView.of(files.map(HierarchicalResource::getPath).collect(Collectors.toList()));
         }