+++ lwb/metalang/sdf3/sdf3/src/main/java/mb/sdf3/task/spec/Sdf3CheckSpec.java
@@ -1,10 +1,10 @@
 package mb.sdf3.task.spec;
 
+import mb.aterm.common.InvalidAstShapeException;
 import mb.common.message.KeyedMessages;
 import mb.common.message.KeyedMessagesBuilder;
 import mb.common.message.Messages;
 import mb.common.message.Severity;
-import mb.common.region.Region;
 import mb.common.result.Result;
 import mb.jsglr.common.JsglrParseException;
 import mb.jsglr.common.JsglrParseOutput;
@@ -14,33 +14,28 @@ import mb.pie.api.ExecContext;
 import mb.pie.api.None;
 import mb.pie.api.TaskDef;
 import mb.pie.api.stamp.resource.ResourceStampers;
-import mb.resource.hierarchical.HierarchicalResource;
 import mb.resource.hierarchical.ResourcePath;
-import mb.resource.hierarchical.match.ResourceMatcher;
-import mb.resource.hierarchical.walk.ResourceWalker;
 import mb.resource.util.SeparatorUtil;
 import mb.sdf3.Sdf3ClassLoaderResources;
 import mb.sdf3.Sdf3Scope;
 import mb.sdf3.task.Sdf3AnalyzeMulti;
+import mb.sdf3.task.Sdf3GetSourceFiles;
 import mb.sdf3.task.Sdf3GetStrategoRuntimeProvider;
 import mb.sdf3.task.Sdf3Parse;
-import mb.sdf3.task.util.Sdf3Util;
-import mb.aterm.common.InvalidAstShapeException;
 import mb.stratego.common.StrategoException;
 import mb.stratego.common.StrategoRuntime;
 import mb.stratego.common.StrategoTermMessageCollector;
-import org.checkerframework.checker.nullness.qual.Nullable;
 import org.spoofax.interpreter.terms.IStrategoTerm;
 import org.spoofax.terms.util.TermUtils;
 
 import javax.inject.Inject;
-import java.util.stream.Stream;
 
 @Sdf3Scope
 public class Sdf3CheckSpec implements TaskDef<Sdf3SpecConfig, KeyedMessages> {
     private final Sdf3ClassLoaderResources classLoaderResources;
     private final Sdf3Parse parse;
     private final Sdf3GetStrategoRuntimeProvider getStrategoRuntimeProvider;
+    private final Sdf3GetSourceFiles getSourceFiles;
     private final Sdf3AnalyzeMulti analyze;
 
     @Inject
@@ -48,11 +43,13 @@ public class Sdf3CheckSpec implements TaskDef<Sdf3SpecConfig, KeyedMessages> {
         Sdf3ClassLoaderResources classLoaderResources,
         Sdf3Parse parse,
         Sdf3GetStrategoRuntimeProvider getStrategoRuntimeProvider,
+        Sdf3GetSourceFiles getSourceFiles,
         Sdf3AnalyzeMulti analyze
     ) {
         this.classLoaderResources = classLoaderResources;
         this.parse = parse;
         this.getStrategoRuntimeProvider = getStrategoRuntimeProvider;
+        this.getSourceFiles = getSourceFiles;
         this.analyze = analyze;
     }
 
@@ -66,23 +63,17 @@ public class Sdf3CheckSpec implements TaskDef<Sdf3SpecConfig, KeyedMessages> {
 
         final StrategoRuntime strategoRuntime = context.require(getStrategoRuntimeProvider, None.instance).getValue().get();
 
-        final ResourceWalker walker = Sdf3Util.createResourceWalker();
-        final ResourceMatcher matcher = Sdf3Util.createResourceMatcher();
-        final HierarchicalResource mainSourceDirectory = context.require(input.mainSourceDirectory, ResourceStampers.modifiedDirRec(walker, matcher));
-        try(final Stream<? extends HierarchicalResource> stream = mainSourceDirectory.walk(walker, matcher)) {
-            final JsglrParseTaskInput.Builder parseInputBuilder = parse.inputBuilder().rootDirectoryHint(input.rootDirectory);
-            stream.forEach(file -> {
-                final ResourcePath filePath = file.getPath();
-                final Result<JsglrParseOutput, JsglrParseException> result = context.require(parseInputBuilder.withFile(filePath).buildSupplier());
-                messagesBuilder.addMessages(filePath, result.mapOrElse(v -> v.messages.asMessages(), e -> e.getOptionalMessages().map(KeyedMessages::asMessages).orElseGet(Messages::of)));
-                result.ifOk(output -> {
-                    checkModuleName(filePath, output.ast, input, messagesBuilder);
-                    collectMessagesFromStrategoStrategy(filePath, output.ast, strategoRuntime, messagesBuilder);
-                });
+        final JsglrParseTaskInput.Builder parseInputBuilder = parse.inputBuilder().rootDirectoryHint(input.rootDirectory);
+        for(ResourcePath file : context.require(getSourceFiles, input.rootDirectory)) {
+            final Result<JsglrParseOutput, JsglrParseException> result = context.require(parseInputBuilder.withFile(file).buildSupplier());
+            messagesBuilder.addMessages(file, result.mapOrElse(v -> v.messages.asMessages(), e -> e.getOptionalMessages().map(KeyedMessages::asMessages).orElseGet(Messages::of)));
+            result.ifOk(output -> {
+                checkModuleName(file, output.ast, input, messagesBuilder);
+                collectMessagesFromStrategoStrategy(file, output.ast, strategoRuntime, messagesBuilder);
             });
         }
 
-        final Sdf3AnalyzeMulti.Input analyzeInput = new Sdf3AnalyzeMulti.Input(input.mainSourceDirectory, parse.createRecoverableMultiAstSupplierFunction(walker, matcher));
+        final Sdf3AnalyzeMulti.Input analyzeInput = new Sdf3AnalyzeMulti.Input(input.rootDirectory, parse.createRecoverableMultiAstSupplierFunction(getSourceFiles.createFunction()));
         final Result<Sdf3AnalyzeMulti.Output, ?> analysisResult = context.require(analyze, analyzeInput);
         analysisResult
             .ifOk(output -> {